Rename ChildBlockHooks to PluginBlockHooks
