+++ plugins/common/src/modules/components/index.js
@@ -1,1 +1,1 @@
-export { default as ChildBlockHooks } from './child-block-hooks';
+export { default as PluginBlockHooks } from './plugin-block-hooks';
